pyproject.toml
[project]
name = "pushward-hass"
version = "0.1.0"
description = "HACS custom integration for Home Assistant — track HA entities as PushWard Live Activities on iPhone"
readme = "CLAUDE.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["custom_components/pushward"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-homeassistant-custom-component>=0.13",
    "aiohttp>=3.9",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
